+++ R/xxx.spatial_write.R
@@ -390,11 +390,11 @@
    if (!nchar(compress))
       return(invisible(NULL))
    if ((.lgrep("gz",compress))&&(nchar(Sys.which("gzip"))))
-      system2("gzip",list(fname))
+      system2("gzip",c(fname))
    else if (.lgrep("bz(ip)*2",compress)&&(nchar(Sys.which("bzip2"))))
-      system2("bzip2",list(fname))
+      system2("bzip2",c(fname))
    else if (.lgrep("xz",compress)&&(nchar(Sys.which("xz"))))
-      system2("xz",list(fname))
+      system2("xz",c(fname))
    else if (compress=="zip") {
       f <- .dir(path=dname
                ,pattern=paste0("^",lname,"\\.",ext,"$")