+++ R/yyy.syno.R
@@ -1,14 +1,6 @@
 '.syn' <- function(new,kind=0,...) { ## function names synonyms
    if (!FALSE) {
-      op <- options(warn=0)
-      n <- sys.parent()-3:1+1
-      n <- n[n>0]
-      res <- as.character(sys.call(which=0))[1]
-      for (i in n) {
-         res <- c(res,as.character(sys.call(which=i))[1])
-      }
-     # print(res)
-      options(op)
+      res <- .parentFunc()
       old <- res[length(res)]
       res <- res[-c(1,length(res))]
       parent <- paste(paste0(res,"()"),collapse=" -> ")
@@ -23,4 +15,15 @@
          message(msg)
    }
    do.call(new,list(...))
+}
+'.parentFunc' <- function() {
+   op <- options(warn=0)
+   n <- sys.parent()-3:1+1
+   n <- n[n>0]
+   res <- as.character(sys.call(which=0))[1]
+   for (i in n) {
+      res <- c(res,as.character(sys.call(which=i))[1])
+   }
+   options(op)
+   res
 }